feat: 마이페이지 첫 화면 SSR 적용
